Test improvements
