+++ src/test/java/sa/com/cloudsolutions/antikythera/evaluator/KitchenSink.java
@@ -1,0 +1,15 @@
+package sa.com.cloudsolutions.antikythera.evaluator;
+
+import java.util.List;
+import java.util.ArrayList;
+import java.util.Map;
+
+public class KitchenSink {
+    private List<String> stringList;
+    private ArrayList<Integer> intList;
+    private Map<Long, Person> people;
+    private String text = "test";
+    private Integer number = 42;
+    private int i = 43;
+
+}
